Major bug fixes

+++ src/converter.js
@@ -268,7 +268,7 @@ const convert = card => {
 
     return {
       property: slot.property.startsWith("Power") ? slot.property === "PowerPoisonAmountAbs" ? `power_poison` : `power_${snakeify(slot.property.slice(5, slot.property.length - 3))}` : slot.property.endsWith("Power") ? "power_unlock" : slot.property === "SP200" ? "power_unlock" : slot.property === "ChargedPowerRegenRate" ? "stat_charged_power_regen" : `stat_${snakeify(slot.property)}`,
-      value: parseInt(slot.value)
+      value: parseFloat(slot.value)
     };
   }
